pyproject.toml
[tool.poetry]
name = "003-isapre"
version = "0.1.0"
description = ""
authors = ["Borja Lozano <borja.lozano@babelgroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
streamlit = "^1.37.0"
langchain = "^0.2.11"
langchain-openai = "^0.1.19"
fpdf = "^1.7.2"
pymupdf = "^1.26.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
